Refactor nearby discovery models and Send screen UI
Move discovery-facing data models into the domain model package so the
NetworkServices repository contract only describes behavior. NearbyDevice now
lives as its own domain model and includes hostAddresses plus port, which makes
the discovered device route explicit for the upcoming local HTTP request/response
work.

Extract Android NsdServiceInfo mapping into a dedicated Android domain extension.
AndroidNetworkServices now delegates resolved NSD service conversion to
toNearbyDeviceAndroidImpl(), keeping the service class focused on registration,
discovery, callback handling, and state updates. The mapper also rejects
incomplete services when required TXT attributes, host addresses, or a valid port
are missing.

Split the Send screen UI into smaller Compose pieces. NearbyDeviceCard and
NearbyDeviceEmptyCard now live under featureSend/components, and SendScreen uses
a shared Sync360Surface wrapper for the rounded surface container. This keeps the
screen focused on collecting ViewModel state and arranging the feature UI.

Add a presentation NearbyDeviceUiModel and domain-to-UI mapper as a first small
step toward keeping UI state separate from domain discovery data.

Update Codex project memory/docs to reflect the current Android-first manual
rebuild, the simplified discovery milestone, host address support, and the next
learning milestone around local HTTP/Ktor request-response.
