+++ _CODEX_CONTEXT_DO_NOT_DELETE/README.md
@@ -1,7 +1,7 @@
-# Codex Project Context
+﻿# Codex Project Context
 
-This folder stores important Codex/AI context for this project.
+This folder stores important Codex/AI context for Sync360.
 
 Do not delete, rename, or move this folder. If the project folder is copied or shared, keep this folder with it.
 
-Future Codex chats should read `_CODEX_CONTEXT_DO_NOT_DELETE/memory.md` before making changes. The memory file is context, not absolute truth: after reading it, inspect the current repository state and update the memory file if major project facts have changed.
+Future Codex chats should read `_CODEX_CONTEXT_DO_NOT_DELETE/memory.md` before working on the project. The memory file is context, not absolute truth: after reading it, inspect the current repository state and update the memory file if major project facts have changed.
