+++ shared/src/commonMain/kotlin/com/liftley/sync360/domain/model/DiscoveryStatus.kt
@@ -1,4 +1,4 @@
-package com.liftley.sync360.domain.local
+package com.liftley.sync360.domain.model
 
 enum class DiscoveryStatus {
     Idle,
