+++ shared/src/commonMain/kotlin/com/liftley/sync360/domain/model/NearbyDevice.kt
@@ -1,0 +1,12 @@
+package com.liftley.sync360.domain.model
+
+data class NearbyDevice(
+    val id: String,
+    val deviceName: String,
+    val deviceType: String,
+    val protocolVersion: String,
+    val hostAddresses: List<String>,
+    val port: Int,
+    val serviceName: String,
+    val serviceType: String
+)
