+++ shared/src/commonMain/kotlin/com/liftley/sync360/presentation/brandComponents/Sync360Surface.kt
@@ -1,0 +1,14 @@
+package com.liftley.sync360.presentation.brandComponents
+
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.runtime.Composable
+
+@Composable
+fun Sync360Surface(content: @Composable (() -> Unit)) {
+    Surface(
+        shape = MaterialTheme.shapes.large
+    ) {
+        content()
+    }
+}
