+++ shared/src/commonMain/kotlin/com/liftley/sync360/presentation/model/NearbyDeviceDomainToUiModel.kt
@@ -1,0 +1,16 @@
+package com.liftley.sync360.presentation.model
+
+import com.liftley.sync360.domain.model.NearbyDevice
+
+fun NearbyDevice.toNearbyDeviceUiModel(): NearbyDeviceUiModel {
+    return NearbyDeviceUiModel(
+        id = this.id,
+        deviceName = this.deviceName,
+        deviceType = this.deviceType,
+        protocolVersion = this.protocolVersion,
+        hostAddresses = this.hostAddresses,
+        port = this.port,
+        serviceName = this.serviceName,
+        serviceType = this.serviceType
+    )
+}
